Attempt to remove unnecessary segmenter configuration
The application was now configured in a way that it wouldn't work at all
without segmenter being configured from .env variable.
It doesn't really seems necessary as the host should always be the same
as the CRM is running at and only people able to access admin should be
able to call CRM/segmenter-related APIs.

This attempts to use AdminLoggedAuthorization to grant access to segmenter
APIs and removes dependency to configure segmenter in .env/configs

remp/crm#581