Logging out users who are not active anymore or their token was removed
Currently if the account is disabled or mass sign-out feature is used,
people's access directly in CRM remains. They should be logged out.

* BasePresenter now emits AuthenticationEvent for others to validate
user's authentication. This ensures that application module doesn't
have any ugly dependency.

* UsersModule is now consistently handling SignIn/Out events and always
creates cookie with access token after login. Before it was split between
UsersModule and private ContentModule. This change is introduced to make
sure that all users can be signed out of CRM when mass sign-out feature
is used.

Mass sign-out cannot be session based and is dependent on some internally
generated and stored token. Access token is that thing.

* User is logged out if the whole account is disabled or if the access
token read from the cookie is not present in the database.

remp/crm#670