Adding new_subscription_delayed event to handle attachments after payment
Process for generation of invoices in CRM is configured to allow users
to enter invoice data before we send them the confirmation email. This
delay allows us to attach them the invoice directly to the confirmation
email.

As there was no nice solution up to now, we emit new_subscription_delayed
event when the subscription is created and attempt to generate attachments
after the configurable delay. Email to Mailer is then sent right away.

remp/crm#890