Fixing issue with default page redirecting user to itself
When CRM is installed, `not_logged_in_route` points to the same
default page as is the main CRM's page. When users tried to click
on admin link, they were directed back to the same page.

It was designed so that ApplicationModule is not dependent on UsersModule's
sign in page.

This commit adds a seeder to UsersModule which overrides default
config and allow users to sign in properly.