Adding `SameSite=Lax` cookie settings to `n_token` cookie
`Strict` wasn't used since it would prevent the cookie to be sent
on requests originated by third parties. Since the cookie is meant
to be used on the CRM and CMS, and articles are meant to be visited
from third parties, it would cause access issues - cookie wouldn't
be sent in these requests at all.

remp/web#1212

Signed-off-by: Peter Dulacka <dulacka@gmail.com>