Fix invalid token source access after source was added to access token DB
Recently we added `access_token.source` to the database. This source got into
conflict with `source` that we used to identify authorization with (such
as Wordpress).

Before, if the Wordpress authenticator got used, it added `source=wordpress`
to the token object and `wordpress` object with extra metadata. After we
added `access_token.source` to the DB, these two got into conflict and
resulted in occasional error in `user/info` API if access token with source
was used to access the API.

remp/dn-mofa#97