Remove redundant scopes in Google Sign in
The openID and profiles scopes were not really necessary
to match user with the Google account. Only email address
is actually required so we lowered the scope requirements.

This lowering was triggered by requirement to update our
privacy policy to describe how we handle personal user data.
We don't need it and we don't want it, hence this change.

remp/novydenik#774