Generate password for Google Sign in registrations
This is temporary solution because login flows and scenarios
are not completely ready yet.

- Scenarios weren't able to trigger "user created" flow because
of missing password. If we enabled it, the password would be
empty in the emails.
- Users would not be able to login other way than via social login.
It's viable to have at least one independent login flow for these
kind of registrations - such as "magic login".
- UserBuilder was reverted to require the password and throw an error
if it wasn't set during the building process.

remp/crm#913