Use custom params validation in UsersEmailHandler
This handler returns 200 OK responses with specific
error codes for some scenarios. The common validation
would cause breaking change for some clients which
was not intended.

Using custom validator returns behavior from pre-1.0
version. We'll be following up with v2 API with
the correct behavior and deprecation of this handler.

Commit also moves email validation to the place where
we already know it doesn't exist in our system to respect
the accompanying comment.

remp/crm#2300