Fix issues with filter registration appearing after `latte/latte` update to version `2.11.5`.
Originally latte searched for filter implementation by searching the registered callback
with lower-cased title of the filter. This version changed that and used original name
of the filter that was used for the registration.

This caused some of our filters to stop working. Main fix is in FilterLoader class.

remp/crm#2511