Disabling resumability for FileResponse downloads
Varnish (since August 21) doesn't allow Content-Range response headers
if the request doesn't contain Range header. Using FileResponse in such
cases ended up with error thrown by Varnish:

    Unexpected content-range header

To prevent these kind of errors, we decided to disable the resumability
for existing FileResponse downloads for now.

This is the commit introducing the change in Varnish:
https://github.com/varnishcache/varnish-cache/commit/4ab11004

We believe this is a bug, because based on the RFC, if the response code
isn't 206 or 416, the Content-Range header should have no meaning.

   The Content-Range header field has no meaning for status codes that
   do not explicitly describe its semantic.  For this specification,
   only the 206 (Partial Content) and 416 (Range Not Satisfiable) status
   codes describe a meaning for Content-Range.

remp/helpdesk#1523